feat(prompt-manager): add chat context support for template prompts #14
This commit introduces the ability to collect and include chat context in template prompts. It modifies the PromptManager to accept a ChatCreationContext and modifies the ChatContextManager to provide a list of ChatContextItems. Additionally, it refactors the GradleTooling to use a more robust method for retrieving Gradle version information and dependencies.