feat(editor-api): enhance QuickAction selection handling #15
Improve the user experience during QuickAction selection by showing a busy state and disabling the quickPick. This prevents the quickPick from hiding immediately and gives feedback to the user that an action is being processed. After the action is executed, the busy state is removed and the quickPick is hidden.