feat(editor): optimize user input selection in commands #4
This commit optimizes the user input selection in the commands file. Now, if a text is selected in the editor, it will be used as input. If no text is selected, the entire document text will be used as input. Also, the import order in the getTheme file has been rearranged.