feat(code-search): add custom TfIdf implementation #19
This commit introduces a custom implementation of the TfIdf class, replacing the previous usage of the "natural" library's TfIdf class. The new implementation includes additional methods and functionality, such as the ability to add documents from a file, list terms, and set custom tokenizers and stopwords. The test file for TfIdfSemanticChunkSearch has been updated to use this new implementation.