feat(ci): add coverage report generation and rename BuiltinCodeSearch #19
This commit adds a new step in the CI workflow to generate a coverage report using the `vitest-coverage-report-action`. This step will always run, even if the tests are failing.

Additionally, the `BuiltinCodeSearch` class has been renamed to `BuiltinFulltextCodeSearch` to better reflect its functionality. This class is designed to perform full-text code searches in a given set of items.