feat(code-search): add OpenTabFilesManager for recent files #19
This commit introduces a new class, OpenTabFilesManager, in the code-search module. This class is responsible for managing the most recent files opened in the editor. It provides methods to get the most recent files based on language ID, extract code snippets from these files, and get the relative paths of these files. The maximum number of relevant files and the snippet length can be configured.