feat(code-parser): add support for Lombok and field parsing in Java code #25
This commit introduces support for parsing Java code that uses the Lombok library. It also adds the ability to parse fields in Java classes. The changes include modifications to the JavaStructurer and JavaLangConfig files to handle Lombok imports and field declarations. A new test case has been added to verify the correct parsing of Lombok and fields in Java code.