refactor(codebase): optimize class and file handling #25
Optimize the handling of classes and files in various modules. Replace the iteration of classes with a Map for better performance in PlantUMLPresenter. Remove unused function in JavaStructurer. Change variable name from 'relatedClasses' to 'codeFiles' in AutoTestActionExecutor for better clarity. Refactor the code in JavaTestGenProvider to use an async function for parsing code files.