feat(code-context): init Go language support in StructurerProvider #24
Added Go language support in StructurerProvider by creating a new GoStructurerProvider class. This class includes a method to retrieve import statements related to the input and output types of a given method. Also, updated the ProviderContainer configuration to include the GoStructurerProvider. Modified the JavaStructurerProvider to use a language ID property instead of a hardcoded string.