feat(search-strategy): add async instruction method and template context
Updated the HydeStrategy interface to include a documentType property and changed the instruction method to return a Promise. Added a KeywordTemplateContext interface and implemented the async instruction method in HydeKeywordsStrategy. Also, added a method in PromptManager to asynchronously retrieve a Hyde template. Enum values in HydeStep and HydeDocumentType have been updated to string literals.