feat(editor-api): add Git log parsing functionality #30
Added a new GitParser module to the editor-api that can parse Git logs into a structured format. This includes the status of the file (added, modified, deleted) and the changes made. Also updated the GitAction module to use this new parser when retrieving changes by hash. Added corresponding tests to verify the correct functionality of the parser.